requirements.txt
asgiref==3.3.4
Django==3.2
mysqlclient==2.0.3
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
structlog==21.1.0
PyMySQL==1.0.2